+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.7 (unreleased)
+1.8 (2023-04-06)
 ----------------
 
 - Allow multiple interfaces to be registered on schedule config.